Handle JSON encoding errors in engine-mpd and show then in UI
The engine-mpd.php will generate a error structure in case an error happens
while encoding/generating the JSON response data (e.g. due to invalid
values which can not be encoded, like NaN or Infinity values).

In the UI this case of error is now checked and handled to show an error
notify to the user. This includes the JSON error message and JSON error code
for easier debugging/fixing.