Update moode.css
Not sure what the sx tag for images relates to.  Think this might make more sense if all of the images for db browsing were tagged with db-browse, but that is currently inconsistent.  If this (or something similar) is appropriate, then a lot of the 'float:left' in the javascript can be removed.