Artist (Year) labels in Tag and Album views
Shows Artist (Year) in Albums column of Tag view, and as last row in Album view.
Hides (0) when Year is not populated in tags, and hides (NaN) when the Year in the tag file is not a number.
Allows correct functionality of Alphabits quick-select bar both normally and when experimental UI.tagViewCovers = false;