Off option and highlight for Recently-Added filter icon
The current 'recently-added' (funnel) filter icon gives no indication as to whether it's active or inactive.
Also, counter-intuitively, clicking the icon turns on the filter, but clicking it again does not turn it off.  The current functionality mixes the selection-reset click for the headers with the click to turn off the filter.

This PR adds:
1. an 'Off' appearance option to hide the Filter icon altogether for those who don't use or need it and prefer a cleaner look.
2. Fills and brightens the Funnel icon when the filter is active - to provide a visual indication.
3. Removes the filter 'turn-off' click from the Genre/Artist and Album headers and makes the Funnel icon a filter turn-on/turn off toggle.