Add h/w to thumb image element
This is supposedly a perf win for native lazy load but didn't seem like much if any so I never bothered with it as it uses some memory but it also mostly maximizes screen use and solves the jitter while scrolling up in the album cover & radio views issue without a placeholder or any extra elements.

This also adds back some missing code from the ill-fated tidal image "fix" pr.