WIP for addressing library lazy load issues
The lazyload library was never designed to deal with dynamically added content so we used settimeout statements to pause for the lists to populate. This doesn't always happen in time which results in no thumbs.

This implements .when().then() to the list output routines with the then() being a call to lazyLode(). This lets the list finish populating before lazyload() is ever called. All other subsequent calls to .lazyload() shouldn't need to be placed in a settimeout statement.

This also:

#1 limits the scope of the .limited class some more

#2 fixes a playlist thumb margin issue

#3 hides the e.g. radio search reset by default so it doesn't flash on/off on page reload

#4 adds object-fit as discussed on the forum

#5 adds lazylode calls to the thumb column setting as when it increases more thumbs are exposed, etc.

#6 reverts the conditional in the setinterval to use coverview & currentView

#7 fixes radio search (span div swap)