Multiple Artists support
Multiple Artists, Performers, and the Conductor of a track are gathered in a common artists array
and now all shown as artist in tag view artist column (when set to Artist or AlbumArtist).
Albums are filtered accordingly when clicking on or more artists and a selected album can be expandend/collapsed.
They also are displayed in each song in the song list.

Added support for albums where AlbumArtists differ in the tracks:
FolderPath key has to be set.
- All AlbumArtists are displayed in the tag view artists column, but only one single album in the album column and in albums view, that is, the album is no longer disrupted.
- When clicking the album, the tracks are shown in the correct order of track numbers.
- When an AlbumArtist is clicked, the album is highlighted and only the tracks with the selected AlbumArtist are displayed.
- Clicking the album expands/collapses the track list.