Change DOMSubtreeModified to MutationObserver
Chrome complains about DOMSubtreeModified so change to use MutationObserver instead.

This also changes the position of the adaptive background preference item so it goes after the controls that get exposed by the #alpha-blend value, changes the genre quick jump index to be like the other columns and puts the whole thing on one <ul><li>...</ul> line for neatness.