[cdsp.php] Fix crash due to empty ['mixers'] array
In the $ymlCfg array the 'mixers' element can be empty for example when a new config is created that does not include any mixers.

Fix is to add an empty() test. We do this for filters, mixers and pipeline.

Also refactored for readability and to keep with coding and variable naming conventions.