[autocfg.php] Fix 'i2sdevice' handler
Should be $_SESSION['i2soverlay'] and not $values['i2soverlay'] because that var doesn't exist in this handler.

This causes $arg to be empty instead if 'autocfg' in the call to cfgI2SDevice($arg) which if the configured audio device is a USB device results in Pi HDMI 1 being configured instead.