[worker.php] Set bluetooth session vars earlier in startup
This fixes an issue where after first boot the param AUDIODEV in file /etc/bluealsaaplay.conf is set = blank instead of "_audioout" in function updDspAndBtInConfs() due to the session var $_SESSION['alsa_output_mode_bt'] not having been set yet in worker.php.

The result is that bluealsa-aplay fails to start because AUDIODEV is blank.